exercise 2 in pattern matching
Back to Elixir. Had to take a break for a few weeks
to brush up on interviewing. Since the Phoenix book
is broken, I'll work on plain old Elixir for a while,
then get back to Phoenix when I have a better idea
how to fix the bug in the example quickly.