p. 55, binary search
This exercise is a bit sneaky. It's an order of magnitude
more involved than any of the previous exercises. That said,
binary search is not terribly difficult, although it's been
something like 20+ years since I implemented one. It's on
my list of things to do in any case, and this just moved it
up the queue.

This commit has a bug which I'm sure will be painfully obvious
once I dig into it, but I'm well over my usual 15 minute time
allotment for daily commit, so I'll poke at it tomorrow or
in the near future.