crawling through the material, busy at work
This morning, as is usual lately, I'm loaded with
things to do at work and anxious to get started.

This commit, as is too usual lately, is another placeholder,
but does represent my commitment to see this through
regardless.

The fiiz buzz example is pretty cool, with private functions
and overloaded pattern. I'll want to give it more time
than I have this morning.