p. 263 another function to duper
This is crawling along, in part because I'm taking on a lot
of responsibility at work, and the material here is actually
pretty complex. At some point, I'm probably going to have to
work my way back through it in a lot more detail, but that's
ok. I'm still getting some "fingerspitzen" or whatever the
word is, for having my fingers learn the ins and outs of
Elixir syntax.